Split up mapping functions by source instrument.
